Use QTypeRevision for all versions and revisions
In many places we carry major and minor versions or revisions that are
loosely coupled to minor versions. As the Qt minor version resets now,
we need to handle these things more systematically. In particular, we
need to add a "major" part to revisions.

QTypeRevision can express the current major/minor pairs more efficiently
and can also be used to add a major version to revisions. This change
does not change the semantics, yet, but only replaces the types.

+++ src/qml/qml/qqmltypemodule_p_p.h
@@ -62,15 +62,15 @@ QT_BEGIN_NAMESPACE
 class QQmlTypeModulePrivate
 {
 public:
-    QQmlTypeModulePrivate(QString module, int majorVersion) :
+    QQmlTypeModulePrivate(QString module, quint8 majorVersion) :
         module(std::move(module)), majorVersion(majorVersion)
     {}
 
     const QString module;
-    const int majorVersion = 0;
+    const quint8 majorVersion = 0;
 
     // Can only ever decrease
-    QAtomicInt minMinorVersion = std::numeric_limits<int>::max();
+    QAtomicInt minMinorVersion = std::numeric_limits<quint8>::max();
 
     // Can only ever increase
     QAtomicInt maxMinorVersion = 0;
